+++ apps/web/src/components/MarketingLayout.tsx
@@ -22,19 +22,20 @@ export function MarketingLayout({
       <NextSeo title={seoTitle} />
       <main
         className={cn(
-          "max-w-screen flex min-h-screen flex-col overflow-x-hidden",
-          isInverted ? "bg-zinc-900 text-gray-200" : "bg-pink-100 text-gray-900"
+          "max-w-screen flex min-h-screen flex-col overflow-x-hidden bg-primary-background text-primary-foreground"
         )}
       >
         <Navbar isInverted={isInverted} />
         <section
           className={clsx(
-            "flex-1",
+            "flex-1 pt-6",
             isMarkdown && "container w-full px-6 py-6 md:px-16"
           )}
         >
           {isMarkdown ? (
-            <div className="prose max-w-full lg:prose-lg">{children}</div>
+            <div className="prose max-w-full dark:prose-invert lg:prose-lg">
+              {children}
+            </div>
           ) : (
             children
           )}