+++ apps/web/src/components/Pricing.tsx
@@ -32,8 +32,7 @@ function PricingElement({
   return (
     <div
       className={cn(
-        "rounded-2xl border border-gray-900 px-4 py-16 lg:border-none",
-        isFeatured && "bg-pink-100",
+        "rounded-2xl border border-accent-background px-4 py-16",
         isFull && "lg:col-span-3 lg:grid lg:grid-cols-3 lg:border-solid lg:py-8"
       )}
     >
@@ -44,9 +43,7 @@ function PricingElement({
         <h2 className={cn("my-6 text-xl font-semibold", isFull && "lg:my-3")}>
           {title}
         </h2>
-        <p className={cn("h-32 text-gray-700", isFull && "lg:h-auto")}>
-          {subtitle}
-        </p>
+        <p className={cn("te h-32", isFull && "lg:h-auto")}>{subtitle}</p>
       </div>
       <div
         className={cn(
@@ -59,9 +56,7 @@ function PricingElement({
           onClick={onClick}
           className={cn(
             "my-6 w-full rounded-xl border px-4 py-2 transition-colors duration-200 ease-in-out",
-            isFeatured || isFull
-              ? "border-pink-300 bg-pink-300 font-medium hover:border-pink-500/70 hover:bg-pink-500/70"
-              : "border-gray-900 hover:bg-pink-100/40",
+            "border-accent-background hover:bg-accent-background",
             isFull && "lg:order-2 lg:w-64 lg:justify-self-center"
           )}
         >
@@ -75,7 +70,7 @@ function PricingElement({
                 <span
                   className={clsx(
                     "flex items-center",
-                    i === 0 ? "font-semibold" : "text-gray-700"
+                    i === 0 ? "font-semibold" : "te"
                   )}
                 >
                   {feature}
@@ -109,7 +104,7 @@ export function PricingTable() {
   const proPlan = getLimitByPlan("PRO");
 
   return (
-    <div className="mx-auto grid max-w-md grid-cols-1 gap-y-4 md:gap-x-4 md:gap-y-12 lg:max-w-none lg:grid-cols-3">
+    <div className="mx-auto grid max-w-md grid-cols-1 gap-y-4 md:gap-x-4 md:gap-y-12 lg:max-w-none lg:grid-cols-4">
       <PricingElement
         onClick={() => signIn("github", { callbackUrl: "/projects" })}
         price="Free"
@@ -161,7 +156,6 @@ export function PricingTable() {
           "Unlimited Feature Flags",
           "Unlimited Environments",
         ]}
-        isFull
       />
     </div>
   );