+++ apps/angular-example/CHANGELOG.md
@@ -1,5 +1,12 @@
 # angular-example
 
+## 0.0.5
+
+### Patch Changes
+
+- @tryabby/angular@1.0.1
+- @tryabby/devtools@4.0.0
+
 ## 0.0.4
 
 ### Patch Changes